+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added setClientIp method to FakeRequest.
 
 ## 1.0.0 - 2018-02-03
 - First stable revision.